What is true or untrue must also be reviewed by you
Wrong reports or "alternative facts" are examples of fake news, i.e. deliberately false reports that spread rapidly on the Internet and social media. Just click on the "share" button and hundreds or even thousands of other users will see it. But how can you recognize Fake News?

That not everything on the Internet is necessarily correct, every user should have noticed in the meantime. With so-called clickbaits, deliberate false reports are often used.

Since fake news is an ever-increasing challenge of our time, we have decided to integrate a fake news filter with this block list into our networks.

## The decision about what is true or untrue must also be reviewed by you